Fix build when http_prefix is configured
This allows us to build the site with absolute URLs (rather than relative)
and have features like search working without post-build edits.
This moves us closer to an automated build.

Builds need to consider the http-prefix because Github pages for repo-specific
sites include the repo name in the first segement of the path, e.g:
https://alphagov.github.io/api-catalogue/

The fork of tech-docs-gem is intended to be temporary, until fixes for issues
documented here have been resolved & merged upstream:
https://github.com/alphagov/api-catalogue/issues/95