Add helper to append value(s) to querystrings
As part of future work, we need to be able to append tracking
information to URLs.

To do this safely, we need to be able to correctly handle the appending
of the querystring element(s) to a potentially existing querystring,
which requires we parse the querystring, then append on any extras
required.

To simplify our tests, we can use a custom matcher,
`have_same_querystring`, so we have more readable tests, and as the
querystring is the important part of our validation.