Add archive for interesting resources and links
To better serve the X-Government API community - in-and-out of our Slack
community - as well as providing some interesting resources for the
private sector, we should publish our interesting links publicly.

This produces both an HTML frontend and an RSS feed, so folks have
options for keeping up-to-date on what's available and being published.

This requires a new CSV data source, and:

- we can utilise the `builder` gem to more easily build up our Atom
  (RSS) feed
- we make sure to add `rel=alternate` to allow for automagic detection
  by feed readers
- add a new `Link` + `Links` class to contain the data required
- add validation through Active Model for the `Link` class, to perform
  batteries-included validation
- add a Rake task to validate that the link data is well-formed, with a
  `schema` namespace + task to make it easier to validate more in the
  future
- add a `/resources/` page as a parent, for any other informative
  resources
- use `rel=alternate` for the original post's link