Update standard to latest version
To address a security fix with standard's transitive dependency
`ansi-regex`, we need to upgrade to the latest version.

Instead of pinning to specific version for standard, which then causes
problems if i.e. Dependabot wants to update the dependencies, we can
instead pin to the major version.

As Standard uses Semantic Versioning, they won't (knowingly) introduce
breaking changes.

We'll still be protected through our lockfile, but have the option of
bumping the dependencies in the future more easily via `npm update`.

As this includes a few major bumps, we've got a number of changes that
need to be automagically applied via `standard --fix`.