Update the link to add an API to the catalogue
During user research we noticed that having to go to the issues page, find
where to create a new issue and then choose the right issue template was
adding friction to the process. This will now take them straight to the page
to create an issue (we don't want to link to the actual issue template as if
we change things like the filename the link would break).