Consume the federated model running in sandbox
We've added methods to retrieve the APIs from the sandbox endpoint, and one to
get APIs from both the csv file and a list of URLs, as we may have multiple
URLs we want to parse in the future.

We're using `Webmock` in the tests to mock the response so that we can unit
test these methods.

We've also added tests to ensure that we have added the `correlation-id` and
accept headers to the request, and that an error is raised if the response has
a 404 or 406 status.

We have created an interface for the `Fetcher`, as although this isn't very
idiomatic in Ruby we think it will be more readable for people coming from
other languages.  Like with the Spring Boot implementation, we've introduced a
`FetcherService` which ensures sure that the fetcher delegates to the correct
version, and throws a `ClientError` if not.

In the `merge` method we could use `map` instead of `each`, but again have opted
for readability over what might be preferred in Ruby.

When creating an `Organisation` we need to provide an `alternative_name` to
link the organisation and API, and and `id` which is used to create the URL to
the entry in the catalogue. We're using the organisation name returned by the
API to create an acronym for the `alternative_name` and a camel-case name for
the `id` to match how organisations are currently represented in tge CSV file.
We might want to re-think how we do this as a separate piece, as there is a
risk of collisions with the `alternative_name` if departments have the same
acronym.