Remove the date fields from the dashboard
The stats on this dashboard are used by us internally, and the last_updated and
first_added fields aren't particularly useful for us.

When we add the ability to fetch APIs from a URL, these stats will be skewed as
it would use the current date, so would skew the results. The simplest solution
to this is to remove these fields from the dashboard.