Add HTMLProofer to validate produced HTML
HTMLProofer is a great tool for catching issues at build time, such as a
broken link, or images that do not contain alt text.

As we've had a few broken links recently, this will catch it sooner,
i.e. in a PR build.

We ignore the hash `#`, as it's commonly used for disabling the Table of
Contents, and in some pages does not render as an anchor.

We also ignore external links, as we shouldn't worry about the integrity
of external links, especially as we may receive intermittent errors from
i.e. a documentation site, and shouldn't be preventing our build when
this happens.