Re-order the API pages layout
During user research we found that the first thing people wanted to see was a
summary of what the API is, it's purpose and what you've got from it, and how
to get authentication. To meet these expectations this moves the description
and licence information to the top of the page.