Update the footer content
We don't want the resources link to be at the top of the navigation bar as it's
not part of the typical user journey, so these have been moved down to the
footer.

We've also removed the `View source` link as this doesn't work as expected and
isn't useful.

The 'Report a problem with this page' link is still in the page content above the footer,
as this is page-specific and we think the footer should contain links that are consistent across
the site.