Add curl to Docker build
We were seeing the following error when running the Docker build: `LoadError:
Could not open library 'libcurl': Error loading shared library libcurl: No such
file or directory.`

The `html-proofer` gem depends on `typhoeus` which in turn depends on
`libcurl`, which needs `curl` to be installed. `curl` isn't automatically
included in Alpine Docker images so we need to add it manually.