Use a json endpoint for the Food Standards Agency
Eventually we want to move to a model where we pull in metadata about the APIs
from an endpoint, rather than having to manually update a CSV file. We're
moving one of the organisations on the catalogue to this format to demonstrate
how it can work with the federated model.

We've ammended the `from_urls` test to mock a 200 response for a request to any
url ending in `apis` so that we don't have to mock every url that we add to our
registries file, as this will get much longer in the future.