Add an endpoint for all apis found with federated model
This means that other catalogues will be able to consume this catalogue using
our federated model of data discovery:
https://github.com/co-cddo/federated-api-model/blob/arnau-handover/design/artefacts/system-context.svg

As this is a static site we don't currently have storage for apis, so this is
slightly different to our model.

Also as this is a static site, to create the 'endpoint' we create a temporary
file that gets copied to the build folder.

For the tests we're comparing the parsed json rather than the raw json strings,
as this allows us to store the expected result in a separate formatted json
file which is more readable, and allows us to compare the actual values.

We're only using this for apis we ingest using the federated model, as we know
these ones follow our schema.