Use our scraper tool to get NHS APIs
We've written a script to scrape data from the NHS catalogue:
https://github.com/co-cddo/api-catalogue-data-source/blob/main/scraper.py

This uses the output of that to update the API catalogue, saving us from
manually updating it.

These APIs won't appear on the stats dasboard any more as the organisation
is created at build time rather than being in the CSV, so the dashboard
stats test has been updated to use DWP's APIs instead.